added images for map editor interface and added hover effect

+++ src/maps/EditorParser.java
@@ -3,7 +3,20 @@ package maps;
 import graph.Graph;
 import graph.GraphNode;
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.HBox;
+import javafx.scene.layout.VBox;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 import menus.MenuHandler;
+import sceneElements.ButtonProperties;
+import sceneElements.Images;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -18,26 +31,66 @@ public class EditorParser {
     private MapEditor editor;
     private String SAVE_DIRECTORY;
     private String SEPERATOR = File.separator;
+    private Stage popUpStage;
+    private boolean overwrite = true;
+    private Image yesImage, noImage, yesImageHover, noImageHover;
 
     public EditorParser(MapEditor mapEditor) {
+
+        setUpPopUp();
+
         String dir = System.getProperty("user.home");
         SAVE_DIRECTORY = dir + SEPERATOR + "bestRTS" + SEPERATOR;
 
         this.editor = mapEditor;
     }
 
-    public void handle(ActionEvent event) {
-        if (event.getSource() == MapEditorInterface.saveButton)
-        {
-            saveToUserFile();
-        }
-        if (event.getSource() == MapEditorInterface.backButton)
-        {
-            MenuHandler.switchScene(MenuHandler.MAIN_MENU);
-        }
+    private void setUpPopUp() {
+        popUpStage = new Stage();
+
+        GridPane messagePanel = new GridPane();
+        VBox layout = new VBox();
+        HBox buttons = new HBox();
+
+        Label message = new Label();
+        Button yes = new Button();
+        Button no = new Button();
+        yesImage = Images.overwriteYes;
+        noImage = Images.overwriteNo;
+        yesImageHover = Images.overwriteYesHovered;
+        noImageHover = Images.overwriteNoHovered;
+
+        ButtonProperties b = new ButtonProperties();
+
+        b.setButtonProperties(yes, "", 0, 0, event -> {
+            overwrite = true;
+            popUpStage.hide();
+        }, new ImageView(yesImage));
+        //b.addHoverEffect2(yes,yesImageHover,yesImage,0,0);
+
+        b.setButtonProperties(no, "", 0, 0, event -> {
+            overwrite = false;
+            popUpStage.hide();
+        }, new ImageView(noImage));
+        //b.addHoverEffect2(no,noImageHover,noImage,0,0);
+
+        buttons.getChildren().addAll(yes, no);
+        layout.getChildren().addAll(message, buttons);
+
+        message.setText("Overwrite existing map?");
+
+        messagePanel.getChildren().addAll(layout);
+        Scene popUpScene = new Scene(messagePanel);
+
+        popUpStage.setScene(popUpScene);
+        popUpStage.initModality(Modality.APPLICATION_MODAL);
+        popUpStage.hide();
+        popUpStage.setOnCloseRequest(event -> {
+            overwrite = false;
+        });
     }
 
-    private void saveToUserFile() {
+    public void saveToUserFile() {
         String fileName = editor.getInterface().getFileName() + ".txt";
         Graph graph = editor.getGraph();
         if (fileName.equals(""))
@@ -58,6 +111,15 @@ public class EditorParser {
                 }
 
                 File savedFile = new File(SAVE_DIRECTORY + fileName);
+
+                if (savedFile.exists())
+                {
+                    popUpStage.showAndWait();
+                }
+
+                if (!overwrite)
+                    return;
+
 
                 FileWriter fileWriter = new FileWriter(savedFile);
